requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.23.2
jinja2==3.1.2
pydantic==2.4.2
websockets==11.0.3 